Implemet Feedback Reply

+++ src/main/java/com/dto/FeedbackDTO.java
@@ -16,7 +16,7 @@ public class FeedbackDTO {
     private String comment;
     private String created_at_time;
     private LocalDate created_date;
-    private int game_id;
+    private long game_id;
     private String g_name;
     private long user_id;
     private String user_name;
